make sure user/group root and bin exist from the start, since they are provided by iso normally
There is an unknown number of spells that depend on user/group bin (and of course, root...).
"bin" is a legacy thing and it's usage shouldn't grow, but we need time to find all spells
still needing it. Until all of them are fixed, I think we need bin account being created
at a very basic level - I chose smgl-fhs because it already severs that purpose for mail, utmp
and such.
Perhaps we think of something better when reworking the accounts stuff and smgl-fhs can be reduced
to only creating dirs.