utils/crm114: correct the build and install
The Makefile has hardcoded CFLAGS, so PRE_BUILD now comments those
before building. Also, the install target of the Makefile is missing a
lot of the useful files, so these are now installed by INSTALL
(includes examples as well as full "programs" like the mailfilter and
mailreaver scripts for spam filtering).