i18n/uim: new spell, universal input method
uim is an input method engine capable of inputting any language
(provided the language module is available, m17n-lib can provide
support for *many* languages), and it can do the input both in X
windows and on the console (via the uim-fep command-line tool). It can
be configured both by hand via a .uim config file, or using a gui
configuration tool.