Revert "Revert "coreutils 7.2""
This reverts commit b40dea18ca1997719e2f1fb683bf300e1ed11f92.

It's not coreutils 7.2 that's failing. Perhaps it's GCC 4.3.3 with the
latest headers on x86?

See Bug #15156

coreutils 7.1 and 7.2 both compile fine with GCC 4.3.3 and glibc 2.9
with default headers on x86_64.
coreutils 7.1 and 7.2 both compile fine with GCC 4.3.3 and glibc 2.9
with 2.6.29 sanitized headers on x86_64.

coreutils 7.1 and 7.2 both fail to compile with GCC 4.3.3 and glibc 2.9
with sanitized headers on x86.

coreutils version does not seem to make any difference.