udev: updated configure options to 143
From NEWS:
The configure options have changed because another library needs to be
installed in a different location. Instead of exec_prefix and udev_prefix,
libdir, rootlibdir and libexecdir are used. The Details are explained in
the README file.

devicekit-disks wasn't able to find udev's pkgconfig files, now it is ok