telephony/asterisk: update to 1.6.1.11
This is a big jump forward update that resulted in a lot of changes. It
works for me, but the spell could still use some improving. For example,
I wanted to get custom build tweaking via make menuconfig working in
BUILD, but I had issues of getting the input/output properly passed back
and forth. I don't have a lot of spare time to work on this, so I'm
pushing it out for other devs to pick up where I've left off. However,
it does work for me, so it's at least usable even if not perfect. :)

I'm pushing the gpg key update as a separate commit, since the asterisk
developers seem to deem publishing their key fingerprints (or even just
the IDs) unnecessary. As a result, I can't properly verify the
trustworthiness of the keys the release was signed with. However, there
was a sha1 hash of the release in the same location as the source and
gpg signature, and I did verify the downloaded source against both the
gpg signature with not-so-trustworthy keys and the sha1 hash.