telephony/asterisk.gpg: add up-to-date keys
Add up-to-date keys for the asterisk developers.

WARNING: THESE KEYS COULD NOT BE PROPERLY VERIFIED!

The most recent release of asterisk was both signed and hashed (sha1),
and I was able to verify the download using both the signatures (using
the keys in this commit) and the sha1. However, I did not have any
channels available for key verification other than downloading from the
public keyserver (pgp.mit.edu).

I am committing these keys now because I have time constraints that will
prevent me from properly verifying these keys within a reasonable time
frame, so hopefully someone else can complete this task sooner. Also, I
am leaving the top-level asterisk.gpg keyring intact in case anyone is
still using it with the other asterisk spells (which should either be
updated or deprecated, accordingly).

If it is found that this commit is in error, then of course it can be
reverted, corrected, and recommitted.