smgl/cauldron: new, Cauldron is now a spell
This spell installs the Cauldron script suite. You can select either the
cauldron scripts (for generating installation media) or the enchantment
sccripts (the run-time installer engine), or both. If enchantment is
installed, then you can also choose which of the installer interfaces
will be installed with it.

The code in this spell is currently less pretty than I would like it to
be, but it works (for now at least), with at least the potential for
easy forward maintainability. Of course improvements are welcome, but
make sure you know what you are doing with it before changing things, as
the cauldron scripts will actually rely on this spell when generating
ISOs.

The script does nothing in BUILD, has some basic depends info (needs to
be fleshed out more), and asks you what you want installed. Other than
that, the brunt of the work is done in INSTALL. Currently only the SCM
version is supported, as there are no Cauldron releases at this time.