firefox: xulrunner dependent regardless of CVS status.
xulrunner was only cast when CVS status was set to FALSE. I found when
cast that plugins under CVS was not working properly. Upon casting
xulrunner manually I found this immediately resolved the issue.

xulrunner has been moved within the DEPENDS file and placed outside
of the CVS IF/ELSE logic so it is called irregardless of CVS status.