init.d/mountall.sh: checkfs() fixed to properly remove /fastboot.
There was an issue in quick /fastboot would not be removed during teh
checkfs() call. The function would return before teh appropriate 'rm'
command was called. In short, if someone placed /fastboot the filesystems
no longer be checked, ever, unless the /fastboot was manually removed
at some point down the road.

This has been corrected. IF logic has been updated to ensure that
/fastboot is removed before the function returns.

This has been tested locally and seems to be working as expeted.