Revert "xorg-server: Updated to 1.9.0"
This reverts commit 8bf80daf9210b0665994cb81879dccb057e0f62c.

All of my machines are on "development" mesalib (7.9) so I missed the
requirement that xorg-server requires 7.8+.

George (p3pilot) caught this:
	configure: error: Package requirements (glproto >= 1.4.10 dri >= 7.8.0) were not met:
	Requested 'dri >= 7.8.0' but version of dri is 7.5.2

The "feature" branch is at 7.7 and I believe we should only be depending
on "stable" branches for our packages, especially something as important
as xorg-server.

Sorry for all those whom've already updated to xorg-server 1.9.0. Know
that downgrading to 1.8.1 worked fine here when I was testing it (still
with mesalib 7.9 ;)).

Note that I did test xorg-server 1.9.0 with xf86-video-ati,
xf86-video-nouveau, and nvidia_driver on three machines for roughly a
month without issue.

Conflicts:

	xorg-xserver/xorg-server/HISTORY