smgl/cauldron: fix configure/install behavior
Corrected the installation of manpages so that they properly install to
subdirectories of /usr/share/man, and only if there is something to
install. Made the query for what installer interfaces to install depend
on whether enchantment was selected to install, since the installer
interfaces are front-ends to enchantment and therefore depend on it.