udev 167
udev 167
========
Bugfixes.

The udev runtime data moved from /dev/.udev/ to /run/udev/. The
/run mountpoint is supposed to be a tmpfs mounted during early boot,
available and writable to for all tools at any time during bootup,
it replaces /var/run/, which should become a symlink some day.

If /run does not exist, or is not writable, udev will fall back using
/dev/.udev/.

On systemd systems with initramfs and LVM used, packagers must
make sure, that the systemd and initramfs versions match. The initramfs
needs to create the /run mountpoint for udev to store the data, and
mount this tmpfs to /run in the rootfs, so the that the udev database
is preserved for the udev version started in the rootfs.

The command 'udevadm info --convert-db' is gone. The udev daemon
itself, at startup, converts any old database version if necessary.

The systemd services files have been reorganized. The udev control
socket is bound by systemd and passed to the started udev daemon.
The udev-settle.service is no longer active by default. Services which
can not handle hotplug setups properly need to actively pull it in, to
act like a barrier. Alternatively the settle service can be
unconditionally
'systemctl'enabled, and act like a barrier for basic.target.

The fstab_import callout is no longer built or installed. Udev
should not be used to mount, does not watch changes to fstab, and
should not mirror fstab values in the udev database.