editors/vim: missing opt. depends, fix patching
There were missing optional dependencies for both Lua and Python3 to be
used as script interpreters.

The initialization of the for loop in PRE_BUILD for processing the patch
files will print bare numbers if the shell doesn't support it (bash
version <4.0, or perhaps if other shells are used), so 1-99 would fail
to match any real filenames. This was corrected by using seq -w
(available in coreutils so everyone has it regardless of shell version),
which pads numbers to the same width using leading 0's.