Dependency Update: Change In Spell Requirements
It was found that without 'extras' in the udev DEPENDS enabled there
were several spells that were needed for the udev spell to cast successfully.
Spell required for udev to cast:
gperf
usbutils
PCIIDS

The spells listed above were moved out of the 'extras' IF block within the
DEPENDS file to insure they are always queried for when udev is cast.

USBIDS was removed from the DEPENDS file. usbutils appears to be explicitly
required for udev to compile properly. Since usbutils provides USBIDS it
does not seem necassary to have both spells listed in the DEPENDS file.