rasqal: remove the broken query language query; It was a needless complication and broke slv2 at runtime unless you selected "all".
Mages can use CONFIG_LOCAL to access that configure choice, and even there one has to grep the configure script itself or just have to know what the correct list of choices is.
So it's very much an insider thing and the default is meant for usage. We want choice _and_ stuff that works. Really, that slv2 breakage (failure to initialize liblrdf) was annoying.