devel/llvm: fix source numbering and gpg for clang
Corrected the source numbering for the clang sources from 2 to 3. When
the gpg source checking was added for the llvm sources, the sigs were
added as SOURCE2, and the clang sources remained as SOURCE2, causing a
conflict if clang was enabled during CONFIGURE, and consequently there
would be a failed download.

Also added upstream gpg checking for the clang sources, akin to the llvm
sources.