devel/llvm: add support for clang static analyzer
Added the option to install the clang static analyzer scripts. The
scripts are functional and helpful for bug-testing code, but are
currently in a beta stage and do not have an official upstream install.
As a result, there is nothing in the makefiles to install the scripts,
nor is there even a recommended/default install location. I have chosed
$INSTALL_ROOT/usr/share/llvm/ as the base directory for where to install
the extra material. This can be changed as necessary, or when there is
an official upstream install location.

The main script(s) (scan-build et. al.) for performing the static
analysis are written in perl, and thus perl is a hard runtime dependency
if the user wishes to install this component. The reports generated by
scan-build (and friends) are indexed html files, and are therefore
viewable using any standard web browser. There is, however, and
additional set of scripts for viewing the reports in scan-view, which
are written in python. As the reports can still be viewed without
running scan-view, python is an optional runtime dependency.