devel/radare2: add devel version
Added the ability to install devel versions via git clone. The official
upstream repository is mercurial, but they maintain an official git
clone. According to the main upstream developer, the git clone tends to
be a bit more stable, as he is willing to push and any all commits to
the mercurial, but then updates the git clone of the mercurial repo when
there is a known degree of stability in the newly pushed patches to the
mercurial repo.

If someone really wants to use the mercurial repo instead of the git
clone, feel free to add that option as well.