chat-im/pidgin: non-optional depends on intltool
CVS support was removed from the spell quite some time ago in DETAILS,
but hadn't been removed from DEPENDS. I removed the CVS if block.

Meanwhile, current pidgin depends on intltool, which was lacking as a
dependency (it was previously inside the CVS if block). Adding intltool
as a regular dependency enables the spell to build correctly.