such major change isn't acceptable for the most number of setups, and won't be acceptable until the sane solution is found globally.
It MUST be done in a branch for the first, then take a discussion
and development process.

I haven't seen /usr on the same partition as root on _good_ setups
for a long time (maybe eight or nine years). And braking this is
more than stupid, especially for those who has at least one
production servers.

We're not Lennarts to brake everything we touch.

Revert "udev: => 177"

This reverts commit bb547a9aed0a8daad21f1ec332105650042aa26b.