Spell Update: aide
The DEPENDS file was updated to include one more requirement that
seems to have been added based upon the documentation available on
the vendors website.

Of further note, zlib does not appear as a dependency (required or
optional) in the vendor documentation. Additionally, aide also
seems to have an optional dependency for PostgreSQL. However, I
have not yet had a chance to test this particular piece of the
functionality.

Aside from the usual items that get updated in DETAILS, there was
only one change that was made which was in regards to the vendor
website. All the references that I have seen have pointed to the
link currently in the updated DETAILS file.