utils/teapot: new spell, advanced spreadsheet
Teapot is an advanced console spreadsheet. It is novel among
spreadsheets in that it gives spreadsheets 3 dimensions instead of the
standard 2, and it is more programmatically logical, referring to cells
as @(x,y,z) instead of A1B2 or $A3$B4 and similar kludges. There is also
a GUI using FLTK.

This spell is currently incomplete but functional. The console client
builds and works fine. The GUI does not currently build, because it
fails to find FLTK since it can't locate the FLTK cmake recipe. I have
FLTK and cmake both installed, but neither seem to install this file,
and I don't know where it comes from.

I also haven't tested building the user manual since it requires lyx and
latex2html. I did at least test that the dependencies are picked up when
selecting the option, however.

Anyone who is more familiar with CMake and can fix the GUI build is
welcome to do so.