Spell Updated: syslog-ng
The DEPENDS file for the spell in question has been updated to more
accurately reflect the dependency around certain libdbi files.

The previous dependancy on 'libdbi' did install the general libdbi
file(s). however, what was still missing were driver files for the
specific database technology that a user may wish to utilize. While
not casting libdbi would certain cause an error in starting syslog-ng
with any form of database implentation, starting it without the
appropriate driver would result in a similar error.

The resolution was to refresh the dependency and replace the current
'libdi' dependancy with 'libdi-drivers'. This allows for the
correct driver(s) to be installed and also for the more general
libdbi libraries and files to be included since 'libdbi-drivers'
is already dependant on 'libdbi'.