pygobject3: added UP_TRIGGERS - cast pygobject if pygobject has gobject-introspection enabled, so it is built without it. See https://bugzilla.gnome.org/show_bug.cgi?id=657054
It is not strictly necessary to for it to be recast since the conflicting files are overwritten by pygobject3, but some files are left over which I *assume* may lead to unintentional consequences. It is recommended that pygobject2 be built without gobject-introspection if pygobect3 is installed in either case.
(cherry picked from commit 518f5051a4a5c7083060e0c12666c83f97fb289c)