crypto/mosh: update to 1.2.2, fix dependencies
Updated the version to 1.2.2, but also fixed some of the dependencies.
The IO:Pty perl module is provided by spell io-tty, so a runtime_depends
was added for this, obsoleting the message in FINAL. Thus, FINAL was
removed. The dependency on libutempter was also made optional, and I
included a link to a discussion on the merits/issues involved with using
libutempter with mosh.

As of mosh 1.2.0, boost is no longer used, so it was removed as a
dependency.