crypto/mosh: direct depends on gcc with CXX
mosh requires a C++ compiler in order to build. Its dependency on
protobuf means a C++ compiler will be accounted for anyway, since
protobuf also requires a C++ compiler to build, but it should still be
listed as a direct dependency of mosh. I forgot to add it in with my
previous commit, but now it's taken care of.