Spell Update: subversion
This is a relatively simple update. In casting this for another
spell, this spell would fail during the ./configure phase of casting
with an error stating that zlib was required in order to compile.

Solution; simply change the DEPENDS file to reflect the requirement.
zlib was changed from an optional dependency to a required one.