glibc: add as_fn_executable_p() to configure        so it can detect grep. I have no idea why or how its missing though.
I found as_fn_executable_p in automake's cache, so I just added it to configure.

The math libs don't compile for me. I'm assuming that it's missing the -lm flags to tell gcc to compile the files with the math lib. But I'm going to assume thats an error with my box for now and fix it later.