lua-forge/luajit: new spell
LuaJIT is a Just-In-Time full implementation of Lua, including both a
compiler and VM. It does not conflict with Lua itself, as it installs to
luajit rather than lua, but it can otherwise act as a drop-in
replacement for Lua 5.1 (with some features from 5.2 that work safely
with 5.1).

There is a compatibility option to enable more 5.2 features, at the
expense of (likely) breaking some 5.1-specific end-user code.