Spell Update: thc-hydra 5.4 -> 7.5
BUILD has been removed as no longer being needed. Based upon the
upstream developer site, thc-hydra only requires a simple
"configure, make, make install" type installation. For this reason
the BUILD file that was there has been removed to allow for the
default build process to be run.

Additionally, details has been updated beyond the general changes
(e.g. - version number, SHA-512 value). While nothing major has been
revised, the SOURCE and SOURCE_URL has been updated. There are three
items that have been noted in the HISTORY file.

As a side note, it does appear that there needs to be a little more
work with dependancies for this spell. In working on updating this
spell and the testing around the update, it seems that the files being
used to validate dependancies and modules needs to be checked.