Spell Update: ntop 3.3.9 -> 5.0.1
Given the time passed since the last update it would stand to
reason to expect a few things to have changed. Do to the length
of time that has passed I wanted to take something of a
"green field" approach and try, to some extent, start fresh.

To that end regular version update changes in DETAILS
notwithstanding, I have removed references to GeoIP within
DETAILS and other files (BUILD, PRE_BUILD, systemgeoip.patch.
This included the removal of the systemgeoip.patch file that
was part of the older version of this spell.

SOURCE_GPG has been replaced with SOURCE_HASH.

BUILD has been removed. Using the default BUILD process produced
not anomolies that showed themselves during my testing and
compiling of this updated version.

Lastly, DEPENDS has been updated. During testing of the new version
it was found that two spells currently listed as required
dependancies (gd and rrdtool) no longer appear needed. The DEPENDS
file has been updated by removing these two files.

As a side note, it appears that the upstream developers have changed
the name of the product/software from "ntop" to "ntopng" and reset
version number. As an example, the latest version of ntop that could
be located is 5.0.1, which this update is made to facilitate.
With the emergance of "ntopng" the number appears to have started
over and is currently sitting at 1.0.

With this change, particularly with the regression of the version
number back to 1.0, would it be simplier to create a new spell for
"ntopng" and deprecate "ntop"?