Spell Deprecation: amap
After this spell was created it was found to already exist under
a similar name, thc-amap. As the duplicate spell was pushed to the
repository, a deprecation was created and the duplicate spell will
be removed after an appropriate amount of time has passed, currently
about two weeks.

Quill was used to setup the deprecation. Besides the minute update
to the HISTORY file, no other changed have been made against quill's
setup.