Spell Deprecation: amap (cont.)
After the initial deprecation attempt for amap it was found with
additional follow up and investigation that the wrong file
(TRIGGERS) was created. This file has been removed and the correct
file (UP_TRIGGERS) has been created with the appropriate lines added.

Also, it was found that in the preferred spell (thc_amap) the
CONFLICTS file was missing. This file has also been created and the
apprpriate line added.