New Spell: gnupg2
This spell is created to differentiate the 'gnupg' spell, currently
at 1.4.14, and a second branch of the same spell, which currently
sits at 2.0.21.

The upstream developer states that these two branch do NOT conflict
with one another and may in fact be used at the same time on the same
machine without adverse consequences.

Dependencies were set using the README file provided in the source
code tarball. It needs to be noted that while the 'libassuan'
library is stated as being required, it did not appear to be needed
when manualling performaing a './configure' and 'make'. I leave it
up to those wider than I to determine if the dependancy requirements
be followed as they stand from the upstream developer, or if the
dependencies be set based upon the spell creator's hands-on
observations.