Spell (nss) Dependancies Removed: sqlite & zip
In testing the spell and various permutations of dependancies, it
was found that sqlite and zip did not appear to be required for
compilation of nss. For this reason, the spells have been removed
from the DEPENDS file.

As the upstream vendor does not mention these programs as either
required or as optional dependencies they are being removed
completely.