Spell Updated: syslog-ng 3.4.5 -> 3.5.1
In addition to the version being bumped up to 3.5.1, the SOURCE_HASH
field was updated from SHA512:WORKS_FOR_ME to SHA1:UPSTREAM_HASH. My
presumption is that using an upstream hash being better than one that
is calculated by developer making the spell update.

balabit.gpg was also removed. DETAILS had some commented out
references to the gpg file in question, but this appears to have been
removed. There no longer appears to be any reference to the gpg file
and for this reason it has been removed.