pkgconfig: Remove glib2 dependency, fixing things to version 0.28.
This duplicates a bit of binary code because pkgconfig now relies on
internal glib2, but this is necessary because glib2 itself depends
on pkgconfig to be built. At least, the pkg-config command should be
working if present. It does not if, because of another bug, the glib2
install fails and the glib2 .so is missing. You cannot install glib2
after that. Pkgconfig is basic enough to warrant some extra care with
external dependencies (which is why they include a copy of glib2 to
begin with, mind).