Revert "openjpeg: => 2.1.0"
This reverts commit d116a6a512a53e1852a6f105974f7588dd641068.

This breaks ffmpeg 2.8.2 (and even the up-to-date 2.8.4).

See https://trac.ffmpeg.org/ticket/2016, the fix doesn't appear to be in
the stable branch yet.

I spent an hour trying to fix ffmpeg to work with openjpeg 2.x, but at
best I got the errors similar to https://trac.ffmpeg.org/ticket/3712
(marked as a duplicate of #2016 above).

Reverting this change allows ffmpeg 2.8.2 (and 2.8.4) to compile (and
run).