CMAKE_CONFIGURE: Default to "Unix Makefiles" for CMAKE_GEN
Error without ninja installed:
CMake was unable to find a build program corresponding to "Ninja".  CMAKE_MAKE_PROGRAM is not set.  You probably need to select a different build tool.

Override CMAKE_GEN based on status of the ninja package. Since we make
this persistent (for subsequent files such as BUILD to see CMAKE_GEN) if
we accept a pre-defined value, then once we've cast with or without
ninja, forever will this dominate our destiny. Or something.