New Spell: pigz
A small, simple spell with two small caveats as noted below:

1) BUILD - there is no "configure" file present. Running 'make' and
its permissible permutations are the only option. As such the default
build behavior would not have worked and a BUILD file was created to
accomodate a simple 'make' command running.

2) INSTALL - the command 'make install' is not possible. To avoid the
need for a patch file to add the missing functionality, I created an
INSTALL file that performs a 'mkdir -p' (likely unneeded) and then
copy the two created files (pigz and unpigz) to the /usr/bin
directory. This was the easiest/simpliest thing I could think of,
however, I am open to ideas of others have a better approach.