Spell Update: etherape 0.9.11 -> 0.9.15
The simple version bump notwithstanding, SECURITY_PATCH was added to
the DETAILS file. There was a potential exploit found in v.0.9.11,
which was subsequently patched in v.0.9.12. Information for the CVE
given below for reference.

CVE-2011-3369
The add_conversation function in conversations.c in EtherApe before
0.9.12 allows remote attackers to cause a denial of service (NULL
pointer dereference and application crash) via an RPC packet, related
to the get_rpc function in decode_proto.c.