net-tools: update to 1.60-git479bb4 to make it build again
Build of old net-tools is broken with newer kernels. Instead of
trying to backport patches, I decided to switch to a git snapshot
from the (again) active development. There are changes in the
output of the commands. That could break things. But SMGL does
not parse ifconfig output, for example. Generally, we rely on
iproute2 anyway, but there are instances of net-tools use that
hopefully are still happy with this updated version.