fftw: switch to separate libraries (upstream default)
This installs the FFTW libraries as separate files, not with a build
combining serial and multithreaded functionality in one library. This
is the usual expected way and fixes software that just counts on
-lfftw3f_threads being available.

A version suffix smgl1 is introduced to enable UP_TRIGGERS to fix spells
that might get broken by the build change. This suffix shall be dropped
on the next version update.