functions for "preservation"
Preservation is the act of migrating the "ownership" of certain files from
an installed spell into a newly created presevation spell. This is
normally used to allow legacy versions of libraries to persist after
dispelling the spell that owns that library. This allows for ABI
breaking upgrades of spells which have the old version of library
removed in the dispel during the upgrade of the spell, leading all
packages that depend on these old versions of the libraries to break.
Once all these dependers are recast they should depend on the new
version of the library instead and the preservation spell can then
safely be removed. When performing security upgrades of libraries it is
important to do all these recasts and remove the preservation spell at
the end, otherwise the system may still be vulnerable.