gnupg-2.2: allow casting of gnupg-2.2 without gpg
When migrating from one gnupg spell to another, such as the migration
from gnupg to gnupg-2.2 which is triggered by the deprecation of gnupg
the source verification of the target spell fails as sorcery attempts to
perform this with gpg, which is not present on the system as the
originally installed spell for gnupg is already removed due to a
conflict from either the original and/or new spell.

This change allow casting of gnupg-2.2 without gpg present on the
system, this is done by doing the source verification manually using
sha512sum from coreutils in PRE_BUILD. Unpacking of the sources is done
using sorcery, but as the external api does not provide a function to do
unpacking without verification this is done by calling into the
internals of sorcery.