qtbase: Disable O_TMPFILE support
The unnamed temporary file support in the QTemporaryFile class interacts
badly with castfs when materializing the file if the target path doesn't
exist, leading to installation errors with "qmake qinstall".

This is probably because it's trying to materialize the file across
filesystem boundaries, maybe due to a bug in castfs.