gcc: drop .la files from installation
This uses the recently introduced infrastructure to influence the
multi-faced build of gcc to avoid libtool archives and also triggers
anything in the system that references things like libstdc++.la.

It worked for me.

The rational for dropping those .la files is that it seems consensus
among GNU/Linux distros to do so and it is just assumed that basic
toolchain libraries from GCC do not appear in libtool form. The pkgsrc
build framework stumbles over this, for example, on an SMGL box. While
one could argue that support and correct handling of /usr/lib/libstc++.la
and friends is a justified bug fix, the other side of the argument is
that nobody else installs those unnecessary libtool archives. They are
implied when you call g++ or gfortran. They really make no sense and
are just clutter anyway.

Convincing upstream to skip generating them would be a perhaps worthwhile
goal, but they may have their reasons and, well, any other distro is capable
of filtering them in the packaging process.