nss: build fix regarding libfreebl
The newer nss builds fail for me because along the line libfreebl.a is not created
and later dearly missed. There seems to be subtle breakage in the build system that
is not commonly found, but happens reliably on my recently updated SMGL system.

My workaround just builds lib/freebl explicitly and then resumes the overall
build if the first attempt fails. Works.