graphics-libs/mesa: fix driver variable usage
Remove leading spaces from all driver list variables before passing them
on to later build scripts. This specifically fixes a bug when a user
selects "all" DRI drivers, since that driver list is initialized to an
empty string before appending drivers in a loop, leading to an
extraneous leading space that, unless culled, gets replaced with a comma
via the BUILD regex. This results in an explicit null option in the
dri-drivers array, which the meson build script treats as an error,
causing the build to fail.

Additionally, make driver messages more consistent. DRI, gallium, and
Vulkan are all different drivers, but the DRI selection remained as
simply selecting which drivers to build (not specifically DRI drivers),
which can be confusing. Now all three driver selections present the same
messages except for replacing the name of the type of drivers to choose.