firefox-esr: version 102.0
This copies the current state of the firefox spell with small modifications,
some of which are to be merged back. In future, the firefox-esr spell should
come into existence as a simple copy of the firefox spell at the appropriate
point in time, only excluding the DETAILS file, which we also might adapt
in future to have an ESR switch in it.