libassuan, libgpg-error, libksba, npth: switch to hash sig to enable upgrade from stable
We need the GPG tools to be installable without GPG tools (dispelled because of conflict
with gnupg-2.4) or with outdated GPG tools (new signatures being incompatible with old
GPG).

In the long run, we should switch to hashes, and maybe optional signatures for users,
but mainly upstream sigs being checked by the grimoire guru on spell update once,
then hash for easy checking on cast being stored.