xscreensaver: fix build for non-PAM systems
The configure script wants to be smart and refuses the build if
it detects traces of PAM on the system. We create these traces
in /etc/pam.d, even with the very cast of xscreensaver, regardless
of linux-pam being present or not.

We could think about the general scheme here, as other spells also install
pam.d files.