net/network-manager: add missing line-continuation
The line continuator (the '\' character) was missing on the
"optional_depends IFCONFIG" line, causing subsequent optional_depends to
not be read correctly and therefore the build took the upstream defaults
for those settings. This could result in build failure if the default
was set to on but the relevant spell wasn't installed in the system, but
will succeed if the user's system happens to have the spells installed.