Revert "nettle: Re-enable castfs, infodir is no reason to disable"
This reverts commit 2f4e22f1efca1c3b89bab583ae87169a672d45b6.

If you want to enable this, find why it happens and fix it. I don't have the time.
Disabling castfs again so that nettle doesn't break gnutls, and then everything that depends on it.

Since this fails during INSTALL, nettle has already been dispelled and leaves my system in a broken state.

Having a functional system is more important than an optional feature like castfs.