dblatex, dblatex3: introduce python3-compatible variant
The new dblatex lives in the same upstream repo, but is a forked branch.
The release tarballs are not announced on the website, but do appear
on sf.net. To play safe, I did not just replace the old dblatex spell and
introduced a provider.

This prompts change of the few spells explicitly depending on dblatex.