exo: add hack to depend on atk variants
This copies from firefox spell the logic to depend on at-spi2-core if
it is recent enough. I don't quite get it, though: Shouldn't we switch
every

	depends atk

to

	depends at-spi2-core

and be done with it? Deprecating atk altogether? Right now, there is a pretty
mess with stuff requesting conflicting atk all the time and failing.