suil-0: dumber syntax as by commit filter rule
I don't like that. As long as we don't have DEPENDS in a DSL, we should
be able to use features of the shell that reduce typos and noise. And even
with a DSL, we really should ensure that we aren't prone to typos writing
enabled, disabled, yes, no, true, false all the time.