build_system_handler.function: new handler
This grimoire function file provides wrappers for the standard "depends"
and "optional_depends" functions that both abstracts across differing
build systems as well as providing the ability to use defaults for any
known and defined build systems. Currently defaults for
autotools/configure (--enable-OPTION_NAME and --disable-OPTION_NAME),
cmake (-DOPTION_NAME=ON and -DOPTION_NAME=OFF), and meson
(OPTION_NAME=enabled and OPTION_NAME=disabled) build systems are
included.

These wrappers are most useful for spells that allow for multiple
differing build systems. Typically this happens because upstream is
transitioning from one build system to another, but some spells may
simply offer support for different build systems as a matter of choice.
However, the functions provided here are also perfectly usable even when
only a single build system is available use with a spell.

The active build system (the one actually being used to build the spell)
is expected to be defined in the environment variable
SPELL_BUILD_SYSTEM. For spells that only have one build system but opt
to use the functions defined here, this variable can be defined in
DETAILS. For spells that offer an option of multiple build systems, this
variable would typically be defined via a form of config_query in
CONFIGURE. The value of SPELL_BUILD_SYSTEM (even if undefined) can be
overridden by use of the -a flag (see below).

The syntax to define the actual dependencies is as follows:

buildsys_depends[_optional] \
                            [-a active_build_system] \
			    [-g grimoire] \
			    [-o global_optname] \
                            spell_name \
			    ...
                            [-b] build_systemN \
			      [-o local_optname] \
			      message \
			      [enabled] \
			      [disabled]
                            [-b] build_systemN+1 \
			      [-o local_optname] \
			      message \
			      [enabled] \
			      [disabled] \
                            ...

Below are examples of how the functions are intended to be used.

DETAILS file:
<code>
...
SPELL_BUILD_SYSTEM=meson
...
</code>

DEPENDS file:
<code>
. /code/smgl/grimoire/build_system_handler.function

buildsys_depends spell1 \
	-b meson \
          -o "mesonopt" \
          message \
	-b configure \
          message \
          --with-foo \
          &&

buildsys_optional_depends -o "foo" spell2 \
	meson\
		message \
		"" \
		"" \
		&&
	configure \
		message \
		&&
...
</code>

The use of the -b flag to indicate the next build system set is optional
if the build system "blocks" can be determined unambiguously. For
example, if surrounding build system blocks are specified with -b, the
one in middle need not be, or if all arguments are explicitly given, or
if it is the last build system block in the set. Note that quoting
arguments to pass them explicitly without giving them a value is
acceptable. However, if the enabled/disabled arguments are elided then
the -b flag is necessary to prevent remaining parameters from being used
to fill the elided arguments of the current build system set.

The -o flag sets the option name. If no -o flag is present then the
spell name is used as the option name. If -o is passed before the spell
name it sets the global option name for that spell. Passing -o within a
build system set overrides the option name locally. The order of
precedence for the option name is:

   local option name > global option name > spell name

The local option name is cleared on entering a new build system set
(i.e. using -b build_system forces a reset of the option name).

Enabled/disabled settings use build system defaults if known unless
explicitly given. For defaults to be known they must be coded into a
function named buildsys_NAME_defaults, where NAME is the build system
name returned by buildsys_mapping. For example, buildsys_cmake_default
or buildsys_autotools_default. Note that explicitly giving the arguments
as empty strings "" does not fully override this behavior. To override
this behavior while also passing an empty string for an enabled/disabled
argument, the local option name must also be explicitly passed as an
empty string.